Fix game component, stick figure, and physics component to address obstacle collision detection

+++ src/menus/Game.java
@@ -179,8 +179,8 @@ public class Game extends JPanel {
         for (Obstacle o: obstacles)
         {
             if (o.canMove()) {
-                int xOrig = o.getxPos();
-                int yOrig = o.getyPos();
+                double xOrig = o.getxPos();
+                double yOrig = o.getyPos();
                 o.updatePosition();
                 if (map.isCollision(o))
                 {
